--- src/main/java/welkit_server/domain/mail/dto/response/EmailResponse.java
@@ -1,12 +1,0 @@
-package welkit_server.domain.mail.dto.response;
-
-import lombok.*;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Builder
-public class EmailResponse {
-
-    private String code;
-}
